LayoutTests/inspector/dom/inspect-expected.txt

Test for DOM.inspect.


== Running test suite: DOM.inspect
-- Running test case: DOM.inspect.valid
PASS: DOMNodeWasInspected event

-- Running test case: DOM.inspect.invalid
PASS: No error
